# Fix Basis `is_orthogonal` and `is_rotation` methods, add `is_orthonormal`