I noticed that the behavior of these methods was not correct for all inputs. This PR adds unit tests for these methods and fixes the methods. If desired, I could expose them and write docs too.

The `is_orthogonal` method was not working for scaled inputs. The new changes also greatly simplify the code, we only need 3 dot products to check if the Basis is orthogonal, because the dot product of orthogonal vectors is zero. Without these changes, these unit tests would fail:

```
./tests/core/math/test_basis.h:340: ERROR: CHECK( Basis::from_scale(Vector3(1.2, 3.4, 5.6)).is_orthogonal() ) is NOT correct!
  values: CHECK( false )
  logged: Basis with only scale should be orthogonal.

./tests/core/math/test_basis.h:348: ERROR: CHECK( Basis(Vector3(3, 4, 0), Vector3(4, -3, 0), Vector3(0, 0, 5)).is_orthogonal() ) is NOT correct!
  values: CHECK( false )
  logged: Basis with a flip, rotation, and uniform scale should be orthogonal.
```

The `is_rotation` method was not taking into account squeeze matrices (a special form of non-uniform scale where the volume is preserved but the scale changes). Admittedly this was an edge case, but still, now it's fixed. Without this fix, this unit test would fail:

```
./tests/core/math/test_basis.h:374: ERROR: CHECK_FALSE( Basis(Vector3(2, 0, 0), Vector3(0, 0.5, 0), Vector3(0, 0, 1)).is_rotation() ) is NOT correct!
  values: CHECK_FALSE( true )
  logged: Basis with a squeeze should not be a rotation.
```

If anyone is wondering, here is the truth table of the Basis validation methods as of this PR:

| Basis has...      | is_rotation | is_orthonormal | is_conformal | is_orthogonal | is_diagonal |
|-------------------|-------------|----------------|--------------|---------------|-------------|
| Rotation          | ✅ true      | ✅ true         | ✅ true       | ✅ true        | ❌ false     |
| Flip              | ❌ false     | ✅ true         | ✅ true       | ✅ true        | ✅ true      |
| Uniform Scale     | ❌ false     | ❌ false        | ✅ true       | ✅ true        | ✅ true      |
| Non-Uniform Scale | ❌ false     | ❌ false        | ❌ false      | ✅ true        | ✅ true      |
| Skew/Shear        | ❌ false     | ❌ false        | ❌ false      | ❌ false       | ❌ false     |
| Nothing/Identity  | ✅ true      | ✅ true         | ✅ true       | ✅ true        | ✅ true      |