@groud 2023-10-13
> The is_orthogonal method was not working for scaled inputs. 

I think that is expected. According to wikipedia:  `In linear algebra, an orthogonal matrix, or orthonormal matrix, is a real square matrix whose columns and rows are orthonormal vectors`. And, still according to wikipedia, "orthonormal vectors" are perpendicular, but they are also supposed to be of length 1. This basically means that a Basis scaling its input is not orthogonal.

I wonder if that's an english specificity though. In french we differentiate "orthonormal" from "orthogonal", where one means unit vectors and the other does not.