@aaronfranke 2023-10-13
@groud Orthogonal and orthonormal are not the same thing. Orthonormal means nornalized orthogonal. Orthogonal does not need to be normalized. We coule add a new method for is_orthonormal. But anyway, I still need a real is_orthogonal method for the glTF transform logic.