@groud 2023-10-13
> @groud Orthogonal and orthonormal are not the same thing.

Well, that's what I mentioned, I was not sure about the definition in English. Here is the article: https://en.wikipedia.org/wiki/Orthogonal_matrix

What I understand from the current `is_orthogonal` method, is that it does in fact returns whether or not the matrix is *orthonormal*. For me I think there's a difference but the wikipedia article about orthogonal/orthonormal matrices says the opposite. 

If those are indeed different, then I think it would be good to add an `is_orthonormal` method for completeness, and fix `is_orthogonal`. This is however a compatibility breaking change that we need to mention. And if they are the same, then this fix is probably wrong. I think we probably need a more reliable source to decide though.