@aaronfranke 2023-10-13
@groud Actually, these methods are not exposed, so it's not a compatibility-breaking change.