@aaronfranke 2023-10-13
I updated the PR to add `is_orthonormal` and I updated the table in the OP.