@lawnjelly 2023-10-16
Orthogonal etymology seems pretty clear:
> "pertaining to or depending upon the use of right angles," 1570s, from French orthogonal, from orthogone, from Late Latin orthogonius, from Greek orthogonios "right-angled," from ortho- "straight" (see ortho-) + gōnia "angle, corner" (from PIE root *genu- (1) "knee; angle"). Related: Orthogonally; orthogonality.

Unfortunately if one influential person or teacher misinterprets a term, it can get regarded as canon. Personally I always go with what the derivation says. :grin: :popcorn: