@dsnopek 2023-12-05
The end state of these changes makes sense to me!

However, if folks have existing code using `is_orthongonal()` which they depend on to also check that there isn't scaling, this change could break that. I don't know how prevalent such code is - I'd personally guess it's pretty rare, but it is technically a breaking change.

The problem with trying to do this in a non-breaking way is the unfortunate naming if we leave `is_orthongonal()` checking if a basis is orthonormal :-)