@aaronfranke 2023-12-05
@dsnopek Note that this method is not exposed to GDScript or GDExtension, so it only affects engine code, or third-party modules.