@groud 2023-12-06
> Barring the question about whether to have the transpose check or the new one for orthonormal I approve in principal. I know @groud has some misgivings, but the etymology is clear. I personally would prioritize correctness over mumbo jumbo. History (hopefully) eventually removes such language mistakes.

I was just worried there would be induced confusion if we did not state clearly (in a comment) which definition of the names we used. But I think it's the right decision to change it, I was just worried about it being a source of confusion for those who would expect "orthogonality" to imply normalization. 

Right now, I think the PR is good in its current state. The comments in the .cpp are enough, and the `.h` has the two function next to each other, so I guess it's clear enough now.