@tagcup3 2023-12-21
@aaronfranke The name "orthogonal" comes from Lie groups. There is no confusion about it, and it is not contested it is only in your mind. An orthogonal matrix M is a representation of the orthogonal group $O(N)$ satisfying the condition $M M^T = I$, see wikipedia https://en.wikipedia.org/wiki/Orthogonal_group or any group theory book.

Please don't change long established mathematical definitions arbitrarily just because you don't know group theory and get confused by it. Ironically, any one who knows what an orthogonal matrix is going to be confused by your implementation.

@reduz @akien-mga  I recommend reverting this commit. It is based on a misunderstand of what orthogonal matrix is.

This commit changes the correct implementation of `is_orthogonal` to an incorrect one, and introduces `is_orthonormal` which behaves like the original `is_orthogonal` but written in an unnecessarily complicated way.