@aaronfranke 2023-12-21
@tagcup3 I apologize, I did some testing, I had a few things mixed up before, some of my statements were not correct.

For `is_orthonormal`, I could not remember any specific case, so tried changing the contents of this method to the old `is_orthogonal` with the transpose-multiply-equals-identity approach and it passes all unit tests. I believe the old code was functioning as an orthonormality check perfectly fine and I was wrong to suggest there was incorrect behavior (even though the method name did not match what it was doing before).

Now, since both produce the correct behavior, we need to figure out which is faster. The code I have in this PR uses fewer approximate equality checks so I would assume it's faster. After benchmarking with looping the test cases one million times, running the old code took 381851 microseconds and the new code took 357442 microseconds, so the code in this PR is 6.4% faster. I also tested with 10 million iterations with random normalized data (1183076µs -> 1093664µs, this PR is 7.5% faster in this case). and random non-normalized data (925739µs -> 820930µs, this PR is 11.3% faster in this case). In all cases the code in this PR is faster. So it's a good improvement.

When I mentioned incorrect behavior what I was thinking about before is the `is_rotation` method, which checks if the determinant is 1, and checks the old `is_orthogonal` method. With the new `is_orthogonal` method this was wrong. I believe I had changed `is_orthogonal` first and then noticed `is_rotation` was wrong and mistakenly come to the conclusion that `is_rotation` was wrong. But anyway, the new logic for `is_rotation` is faster.

As for `is_orthogonal` vs `is_perpendicular` vs `is_rotation_and_scale`, all seem fine to me. The last one seems fairly cumbersome of a name but it is highly accurate and describes its use case. I disagree with your assessment that `is_perpendicular` is unsuitable, it is likewise descriptive to me.

TL;DR: Everything is good in master except for bikeshedding about the naming of this method. If anyone proposes to change it to `is_perpendicular` or `is_rotation_and_scale` I am fine with that. The most important thing is that this method must exist because I need it for figuring out if a Basis is decomposable into rotation and scale.